+++ test/namedtuple.jl
@@ -21,6 +21,10 @@
 @test (a=3,)[:a] == 3
 @test (x=4, y=5, z=6).y == 5
 @test (x=4, y=5, z=6).z == 6
+@test (x=4, y=5, z=6)[1:1] == (x=4,)
+@test (x=4, y=5, z=6)[1:2] == (x=4, y=5)
+@test (x=4, y=5, z=6)[[1,3]] == (x=4, z=6)
+@test (x=4, y=5, z=6)[[:x,:y]] == (x=4, y=5)
 @test_throws ErrorException (x=4, y=5, z=6).a
 @test_throws BoundsError (a=2,)[0]
 @test_throws BoundsError (a=2,)[2]